Relax criteria on hits
Improved the naive screen-space ray tracing implementation by relaxing
the criteria on hits. Counterintuitively, this imrpoves visual quality
in spite of decreasing precision. The reason for this is that artifacts
that were cause by no hits previously are filled now. These imprecise
fills are still a lot better than outright artifacts.

As this patch reduces artifacts, reinstated the bunny into the demo and enabled
specular reflectance for the floor.