Add binary search for ray marching refinement
Ray marching takes large strides (16 pixels) accross the image now and
the final match is refined by a binary search-like procedure.

Added new criteria for matches so that in addition to z-coordinates, the
direction of the reflected ray and the direction between the reflected
and reflector fragments must also closely match. The binary search-like
procedure performing also converges on ray direction.

Changed depth testing to work with reciprocal z-coordinates so that
depth values can be linearly interpolated in view space.