Move types and functions to modules
Types and functions used by the screen-space reflections shader are
moved to their own separate files. They are included in the shader
using the 'glslify-import' transform that allows including code from
other files without changing names. This is important, because two
functions use user-defined types as return values/arguments, and
would break because of the renaming if they were included simply with
'glslify'.