Fix checking for occluder
If the ray march encounters an object in screen space that lies along
the reflected ray in view space, but it is closer to the camera than the
end of the reflected ray at the current iteration, then that object
occludes possible later hits, so the ray march should be terminated.
Previously, this was done by checking whether the z coordinate of the
encountered object is within some margin, but this approach allowed
false positives. Now we check explicitly if the object is along the
reflected ray.