Hack to fix curved objects' bottoms
This is a quick hack to fix the artifacts visible at the bottom of
reflecting curved objects. The ray march fails on these surfaces,
because it steps over them and because they can reflect objects closer
to the camera than they are. The problem with rays heading towards the
camera is that they are likely to travel out of the frame, encounter an
occluder and the depth test can't handle them.

To provide a quick fix for these issues, rays that travel towards the
camera are now killed and the tolerance for direction matching is
lowered for down-facing surfaces to allow for coarse matches that can
then be refined by binary search. The assumption is that down-facing
surfaces are close to the ground and a higher variance is tolerable for
short distances.